# api/openapi.yaml
openapi: 3.0.3
info:
  title: Blog API
  description: |
    RESTful API for multi-tenant blog platform.

    ## Authentication

    All endpoints (except `/api/health`) require API key authentication.

    Include your API key in one of these headers:
    - `Authorization: Bearer <your-api-key>`
    - `X-API-Key: <your-api-key>`

    ## API Key Types

    - **Admin keys** (`sa_`): Full access to all resources
    - **User keys** (`sk_`): Access to owned/member sites
    - **Site keys** (`ss_`): Access limited to specific site

    ## Scopes

    - `read`: Read-only access
    - `write`: Create and update resources
    - `delete`: Delete resources
    - `admin`: Full administrative access
  version: 1.0.0
  contact:
    name: API Support
  license:
    name: MIT
servers:
  - url: http://localhost:3001
    description: Development server
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      description: API key as Bearer token
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key in X-API-Key header
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
          description: Error type
        message:
          type: string
          description: Error details
      required:
        - error
    Pagination:
      type: object
      properties:
        page:
          type: integer
          minimum: 1
        limit:
          type: integer
          minimum: 1
          maximum: 100
        total:
          type: integer
        totalPages:
          type: integer
...
    Post:
      type: object
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
        slug:
          type: string
        description:
          type: string
          nullable: true
        excerpt:
          type: string
          nullable: true
        content:
          type: string
          nullable: true
        content_format:
          type: string
          enum: [markdown, html, plaintext]
        status:
          type: string
          enum: [draft, published, archived, scheduled]
        featured_image_url:
          type: string
          nullable: true
        author_id:
          type: string
          format: uuid
        site_id:
          type: string
          format: uuid
        series_id:
          type: string
          format: uuid
          nullable: true
        series_part:
          type: integer
          nullable: true
        word_count:
          type: integer
        reading_time_minutes:
          type: integer
        published_at:
          type: string
          format: date-time
          nullable: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    PostCreate:
      type: object
      required:
        - title
        - slug
      properties:
        title:
          type: string
          minLength: 1
        slug:
          type: string
          pattern: '^[a-z0-9]+(-[a-z0-9]+)*$'
        description:
          type: string
        excerpt:
          type: string
        content:
          type: string
        content_format:
          type: string
          enum: [markdown, html, plaintext]
          default: markdown
        status:
          type: string
          enum: [draft, published, archived, scheduled]
          default: draft
        featured_image_url:
          type: string
        author_id:
          type: string
          format: uuid
          description: Only for admin keys
        series_id:
          type: string
          format: uuid
        series_part:
          type: integer
        category_ids:
          type: array
          items:
            type: string
            format: uuid
        tag_ids:
          type: array
          items:
            type: string
            format: uuid
        seo:
          $ref: '#/components/schemas/PostSEO'
    PostUpdate:
      type: object
      properties:
        title:
          type: string
        slug:
          type: string
        description:
          type: string
        excerpt:
          type: string
        content:
          type: string
        content_format:
          type: string
          enum: [markdown, html, plaintext]
        status:
          type: string
          enum: [draft, published, archived, scheduled]
        featured_image_url:
          type: string
        series_id:
          type: string
          format: uuid
        series_part:
          type: integer
        category_ids:
          type: array
          items:
            type: string
            format: uuid
        tag_ids:
          type: array
          items:
            type: string
            format: uuid
        seo:
          $ref: '#/components/schemas/PostSEO'
    PostSEO:
      type: object
      properties:
        meta_title:
          type: string
        meta_description:
          type: string
        focus_keyword:
          type: string
        canonical_url:
          type: string
        og_title:
          type: string
        og_description:
          type: string
        og_image:
          type: string
...
security:
  - BearerAuth: []
  - ApiKeyAuth: []
paths:
...
  /api/sites/{siteId}/posts:
    parameters:
      - name: siteId
        in: path
        required: true
        schema:
          type: string
          format: uuid

    get:
      tags:
        - Posts
      summary: List posts
      description: List all posts for a site with filtering and pagination
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [draft, published, archived, scheduled]
        - name: author_id
          in: query
          schema:
            type: string
            format: uuid
        - name: category_id
          in: query
          schema:
            type: string
            format: uuid
        - name: tag_id
          in: query
          schema:
            type: string
            format: uuid
        - name: search
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
            maximum: 100
        - name: sort
          in: query
          schema:
            type: string
            default: created_at
        - name: order
          in: query
          schema:
            type: string
            enum: [asc, desc]
            default: desc
      responses:
        '200':
          description: List of posts
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Post'
                  pagination:
                    $ref: '#/components/schemas/Pagination'

    post:
      tags:
        - Posts
      summary: Create post
      description: Create a new post
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PostCreate'
      responses:
        '201':
          description: Post created
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/Post'
        '409':
          description: Conflict - slug already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/posts/{postId}:
    parameters:
      - name: postId
        in: path
        required: true
        schema:
          type: string
          format: uuid

    get:
      tags:
        - Posts
      summary: Get post
      description: Get a specific post by ID
      responses:
        '200':
          description: Post details
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/Post'
        '404':
          description: Post not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    patch:
      tags:
        - Posts
      summary: Update post
      description: Update a post
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PostUpdate'
      responses:
        '200':
          description: Post updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/Post'
        '404':
          description: Post not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    delete:
      tags:
        - Posts
      summary: Delete post
      description: Delete a post (requires delete scope)
      responses:
        '204':
          description: Post deleted
        '404':
          description: Post not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  # Categories
...